backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Xiang Gao <xgao4679@usc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
llvmlite = "^0.44.0"
numba = "^0.61.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
sentence-transformers = "^3.4.1"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
